+++ gw_full_latest/gw_web_admin/admin/libs/python/key_thingspeak.py
@@ -8,6 +8,7 @@ import key_GpsFile
 import key_MQTT
 import key_NodeRed
 import key_TTN
+import key_ChirpStack
 
 key_data = {'thingspeak_channel_key':key_ThingSpeak._def_thingspeak_channel_key,
 		'thingspeak_source_list':key_ThingSpeak.source_list,
@@ -29,8 +30,9 @@ key_data = {'thingspeak_channel_key':key_ThingSpeak._def_thingspeak_channel_key,
 		'nodered_sensor_name':key_NodeRed.sensor_name,
 		'nodered_source_list':key_NodeRed.source_list,
 		'nointernet_source_list':key_NoInternet.source_list,
-		#'ttn_server':key_TTN.TTN_server,
-		#'ttn_port':key_TTN.TTN_port,		
-		'ttn_source_list':key_TTN.source_list
+		#'ttn_server':key_TTN.lorawan_server,
+		#'ttn_port':key_TTN.lorawan_port,		
+		'ttn_source_list':key_TTN.source_list,
+		'chirpstack_source_list':key_ChirpStack.source_list
 		}
 print json.dumps(key_data)