+++ gw_full_latest/gw_web_admin/admin/pages/cloudGpsFile.php
@@ -111,7 +111,7 @@
 				</td>			
 			</tr>
 			<tr>
-				<td>source list</td>
+				<td><a href="#" class="my_tooltip" data-toggle="tooltip" title="Indicate a list of accepted device addresses, e.g. 6,7,0x01020304">source list</a></td>
 				<td id="cloudGpsFile_source_list_value">
 					<?php 
 						$array = $key_clouds['gpsfile_source_list'];
@@ -132,12 +132,11 @@
 				<td id="td_edit_cloudGpsFile_source_list">
 					<div id="div_cloudGpsFile_source_list" class="form-group">
 					    		<label>Add a sensor to your list</label>
-                                			<input id="cloudGpsFile_source_list_input" class="form-control" placeholder="e.g. 6,7,8" type="text" 
+                                			<input id="cloudGpsFile_source_list_input" class="form-control" placeholder="e.g. 6,7,0x01020304" type="text" 
 							value="<?php 
 								$array = $key_clouds['gpsfile_source_list'];
 								$size = count($key_clouds['gpsfile_source_list']);
 								$i = 0;
-								if($size == 0) echo "Empty";
 								foreach( $array as &$cel){
 									if(($size -1) == $i ){
 										echo $cel ;