+++ gw_full_latest/key_ThingSpeak.py
@@ -4,7 +4,7 @@ _def_thingspeak_channel_key='SGSH52UGPVAUYG3S'
 #Note how we can indicate a device source addr that are allowed to use the script
 #Use decimal between 2-255 and use 4-byte hex format for LoRaWAN devAddr
 #leave empty to allow all devices
-#source_list=["6", "7", "01020304"]
+#source_list=["6", "7", "0x01020304"]
 source_list=[]
 
 key_association=[]