+++ gw_full_latest/ChangeLog.md
@@ -3,10 +3,13 @@ Change logs
 
 Dec 24th, 2019
 --------------
-- when using a RAK831/2245 SX1301-based hat with our `lora_pkt_fwd_formatter.py` formatter script to provide multi-channel support, we tested that a LoRaWAN device (commercial or LMIC-based) is fully supported (uplink, downlink, in both ABP and OTAA mode).
+- when using a RAK831/2245 SX1301-based hat with our `lora_pkt_fwd_formatter.py` formatter script to provide multi-channel support using a modified version of Semtech's `lora_pkt_fwd`, we tested that a LoRaWAN device (commercial or LMIC-based) is fully supported (uplink, downlink, in both ABP and OTAA mode).
+- `scripts/ttn` has been renamed to `scripts/lorawan_stats`
 - the gateway can also interact with the ChirpStack open-source LoRaWAN Network Server. Tests have been realized with a locally installed ChirpStack NS (127.0.0.1).
 	* `CloudChirpStack.py`, `key_ChirpStack.py` and `scripts/lorawan_stats/cs_stats.py` are added. They are very similar to the TTN scripts but we choose to duplicate them so that both TTN and ChirpStack can be used at the same time if this is desirable.
+	* `scripts/chirpstack` has been added and contains shell scripts to start/stop ChirpStack
 	* see `scripts/chirpstack/README.md` for more details
+- the gateway web admin interface has been improved and also includes support of ChirpStack	
 
 
 Dec 12th, 2019
@@ -92,7 +95,7 @@ April 25th, 2019
 
 March 26rd, 2019
 ----------------
-- add support of TheThingsNetwork (TTN): `CloudTTN.py`, `key_TTN.py` and `scripts/lorawan_stats/ttn_stats.py` are added
+- add support of TheThingsNetwork (TTN): `CloudTTN.py`, `key_TTN.py` and `scripts/ttn/ttn_stats.py` are added
 - the regular gateway id has been changed from 00000027EBD1B236 (for instance) to 0000B827EBD1B236 to take the 6 bytes of the eth0 MAC address
 - the gateway id for TTN will then be defined as B827EBFFFFD1B236
 - there is a dedicated README file for TTN support and configuration